Version 0.0.3
Cleaned up signup command, added 4 user limit for signup. temp.

+++ commands/Hunger Games/signup.js
@@ -4,6 +4,7 @@ const fs = require('fs');
 const p1 = "D:/Hunger Games Bot/Data/SignupUsers/user.txt"
 const p2 = "D:/Hunger Games Bot/Data/SignupUsers/user2.txt"
 const p3 = "D:/Hunger Games Bot/Data/SignupUsers/user3.txt"
+const p4 = "D:/Hunger Games Bot/Data/SignupUsers/user4.txt"
 const {
     Client,
     RichEmbed
@@ -22,10 +23,21 @@ class signup extends commando.Command {
         async run(message, args) {
           var user = message.author.username;
 
-          message.channel.send(user);
-          if (fs.existsSync(p1)) {
-              if (fs.existsSync(p2)) {
-                fs.writeFile("D:/Hunger Games Bot/Data/SignupUsers/user3.txt", user, {
+          if (fs.existsSync(p4)) {
+            message.reply("No more signups at this time");
+          }
+          else if(fs.existsSync(p3)){
+            fs.writeFile("D:/Hunger Games Bot/Data/SignupUsers/user4.txt", user, {
+            encoding: "utf-8"
+            }, function(err) {
+            if (err) {
+            throw err;
+            }
+            console.log("file user4 saved!");
+            });
+          }
+          else if (fs.existsSync(p2)) {
+            fs.writeFile("D:/Hunger Games Bot/Data/SignupUsers/user3.txt", user, {
             encoding: "utf-8"
             }, function(err) {
             if (err) {
@@ -33,30 +45,40 @@ class signup extends commando.Command {
             }
             console.log("file user3 saved!");
             });
-              }
-              else{
+          }
+          else if (fs.existsSync(p1)){
             fs.writeFile("D:/Hunger Games Bot/Data/SignupUsers/user2.txt", user, {
-        encoding: "utf-8"
-        }, function(err) {
-        if (err) {
-        throw err;
-        }
-        console.log("file user2 saved!");
-        });
-  }
+            encoding: "utf-8"
+            }, function(err) {
+            if (err) {
+            throw err;
+            }
+            console.log("file user2 saved!");
+            });
+          }
+          else{
+            fs.writeFile("D:/Hunger Games Bot/Data/SignupUsers/user.txt", user, {
+            encoding: "utf-8"
+            }, function(err) {
+            if (err) {
+            throw err;
+            }
+            console.log("file user saved!");
+            });
+          }
+
 }
-else{
-          fs.writeFile("D:/Hunger Games Bot/Data/SignupUsers/user.txt", user, {
-  encoding: "utf-8"
-}, function(err) {
-  if (err) {
-    throw err;
-  }
-  console.log("file user saved!");
-});
-message.channel.send("debug");
-        }
-}
-}
+
+          }
+
+
+
+
+
+
+
+
+
+
 
 module.exports = signup;
